requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.1.12
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djongo==1.3.6
drf-yasg==1.21.7
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
packaging==23.2
Pillow==10.1.0
pluggy==1.3.0
PyJWT==2.8.0
pymongo==3.12.3
pyTelegramBotAPI==4.14.0
pytest==7.4.3
pytest-django==4.6.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.2.4
telebot==0.0.5
uritemplate==4.1.1
urllib3==2.0.7
channels==3.0.4
